tests/test_routes.py
"""
Test cases for Product API Service

Test cases can be run with:
    nosetests
  
"""

from crypt import methods
import json
import os
import logging
import unittest

# from unittest.mock import MagicMock, patch
from service.utils import status  # HTTP Status Codes
from service.models import db, init_db
from service.routes import app
from tests.factories import ProductFactory
from service.models import Product


# Disable all but critical errors during normal test run
# uncomment for debugging failing tests
logging.disable(logging.CRITICAL)

# DATABASE_URI = os.getenv('DATABASE_URI', 'sqlite:///../db/test.db')
DATABASE_URI = os.getenv(
    "DATABASE_URI", "postgresql://postgres:postgres@localhost:5432/testdb"
)

BASE_URL = "/products" # change
CONTENT_TYPE_JSON = "application/json"


######################################################################
#  T E S T   C A S E S
######################################################################
class TestProductServer(unittest.TestCase):
    """Test Cases for Product Service"""

    @classmethod
    def setUpClass(cls):
        """Run once before all tests"""
        app.config["TESTING"] = True
        app.config["DEBUG"] = False
        # Set up the test database
        app.config["SQLALCHEMY_DATABASE_URI"] = DATABASE_URI
        app.logger.setLevel(logging.CRITICAL)
        init_db(app)

    @classmethod
    def tearDownClass(cls):
        """Run once after all tests"""
        db.session.close()

    def setUp(self):
        """Runs before each test"""
        db.drop_all()  # clean up the last tests
        db.create_all()  # create new tables
        self.app = app.test_client()

    def tearDown(self):
        db.session.remove()
        db.drop_all()

    def _create_products(self, count):
        """Factory method to create products in bulk"""
        products = []
        for _ in range(count):
            test_product = ProductFactory()
            resp = self.app.post(
                BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
            )
            self.assertEqual(
                resp.status_code, status.HTTP_201_CREATED, "Could not create test product"
            )
            new_product = resp.get_json()
            test_product.id = new_product["id"]
            products.append(test_product)
        return products

    def test_index(self):
        """Test the Home Page"""
        resp = self.app.get("/")
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        #data = resp.get_json()
        #print(data)
        #self.assertEqual(data["name"], "Product REST API Service")

    def test_get_product_list(self):
        """Get a list of Products"""
        self._create_products(5)
        resp = self.app.get(BASE_URL)
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        data = resp.get_json()
        self.assertEqual(len(data), 5)

    def test_get_product(self):
        """Get a single Product"""
        # get the id of a product
        test_product = self._create_products(1)[0]
        resp = self.app.get(
            "/products/{}".format(test_product.id), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        data = resp.get_json()
        self.assertEqual(data["name"], test_product.name)

    def test_get_product_with_name(self):
        """Query Products by name"""
        test_product = ProductFactory()
        logging.debug(test_product)
        resp = self.app.post(
            BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)

        test_product = resp.get_json()
        # resp = self.app.get(
        #     "/products?name={}".format(test_product['name']), content_type=CONTENT_TYPE_JSON
        # )
        resp = self.app.get(
            BASE_URL,
            query_string="name={}".format(test_product['name'])
        )
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        data = resp.get_json()
        self.assertEqual(data[0]["name"], test_product['name'])
        self.assertEqual(data[0]["category"], test_product['category'])
        self.assertEqual(data[0]["description"], test_product['description'])
        self.assertEqual(data[0]["price"], test_product['price'])
        self.assertEqual(data[0]["stock"], test_product['stock'])


    def test_get_product_with_category(self):
        """Query Products by category"""
        test_product = ProductFactory()
        logging.debug(test_product)
        resp = self.app.post(
            BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)

        test_product = resp.get_json()
        # resp = self.app.get(
        #     "/products?category={}".format(test_product['category']), content_type=CONTENT_TYPE_JSON
        # )
        resp = self.app.get(
            BASE_URL,
            query_string="category={}".format(test_product['category'])
        )
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        data = resp.get_json()
        self.assertEqual(data[0]["name"], test_product['name'])
        self.assertEqual(data[0]["category"], test_product['category'])
        self.assertEqual(data[0]["description"], test_product['description'])
        self.assertEqual(data[0]["price"], test_product['price'])
        self.assertEqual(data[0]["stock"], test_product['stock'])

    def test_create_product(self):
        """Create a new Product"""
        test_product = ProductFactory()
        logging.debug(test_product)
        resp = self.app.post(
            BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)
        # Make sure location header is set
        location = resp.headers.get("Location", None)
        self.assertIsNotNone(location)
        # Check the data is correct
        new_product = resp.get_json()
        self.assertEqual(new_product["name"], test_product.name, "Names do not match")
        self.assertEqual(
            new_product["description"], test_product.description, "Description do not match"
        )
        self.assertEqual(
            new_product["price"], test_product.price, "Price does not match"
        )
        # Check that the location header was correct
        resp = self.app.get(location, content_type=CONTENT_TYPE_JSON)
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        new_product = resp.get_json()
        self.assertEqual(new_product["name"], test_product.name, "Names do not match")
        self.assertEqual(
            new_product["description"], test_product.description, "Description do not match"
        )
        self.assertEqual(
            new_product["price"], test_product.price, "Price does not match"
        )

    def test_update_product(self):
        """Update an existing Product"""
        # create a product to update
        test_product = ProductFactory()
        resp = self.app.post(
            BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)

        # update the product
        new_product = resp.get_json()
        logging.debug(new_product)
        new_product["name"] = "Huawei"
        resp = self.app.put(
            "/products/{}".format(new_product["id"]),
            json=new_product,
            content_type=CONTENT_TYPE_JSON,
        )
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        updated_product = resp.get_json()
        self.assertEqual(updated_product["name"], "Huawei")

    def test_update_product_nothing(self):
        """Update no-existing Product"""
        resp = self.app.put(
            "/products/{}".format(0),
            json={},
            content_type=CONTENT_TYPE_JSON,
        )
        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)


    def test_delete_product(self):
        """Delete a Product"""
        test_product = self._create_products(1)[0]
        resp = self.app.delete(
            "{0}/{1}".format(BASE_URL, test_product.id), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_204_NO_CONTENT)
        self.assertEqual(len(resp.data), 0)
        # make sure they are deleted
        resp = self.app.get(
            "{0}/{1}".format(BASE_URL, test_product.id), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)

    def test_delete_product_no_data(self):
        """Delete a Product with no data"""
        resp = self.app.delete(
            "{0}/{1}".format(BASE_URL, 0), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)

    def test_create_product_no_data(self):
        """Create a Product with missing data"""
        resp = self.app.post(BASE_URL, json={}, content_type=CONTENT_TYPE_JSON)
        self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)

    def test_create_product_bad_price_string(self):
        """ Create a Product with bad available data"""
        test_product = ProductFactory()
        logging.debug(test_product)
        # change price to a string
        test_product.price = "dollar"
        resp = self.app.post(
            BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
    
    #def test_create_product_bad_price_negative(self):
    #    """ Create a Product with bad available data"""
    #    test_product = ProductFactory()
    #    logging.debug(test_product)
    #    # change available to a string
    #    test_product.price = -1000
    #    resp = self.app.post(
    #        BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
    #    )
    #    self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)

    def test_get_product_not_found(self):
        """Get a Product that not found"""
        resp = self.app.get("/products/0")
        self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)

    def test_method_not_allowed(self):
        """Send a request with a method that is not allowed"""
        resp = self.app.delete(BASE_URL)
        self.assertEqual(resp.status_code, status.HTTP_405_METHOD_NOT_ALLOWED)
    
    def test_create_product_no_content_type(self):
        """Create a Product with no content type"""
        resp = self.app.post(BASE_URL)
        self.assertEqual(resp.status_code, status.HTTP_415_UNSUPPORTED_MEDIA_TYPE)

    def test_healthcheck(self):
        """Check healthcheck function"""
        resp = self.app.get("/healthcheck")
        data = resp.get_json()
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        self.assertEqual(data['message'], "Healthy")
    
    def test_purchase_product(self):
        """Purchase a Product"""
        # create a product to purchase
        test_product = Product(name="Xiaomi", category="Phone", description="Test for purchase", price=999, stock=5)
        resp = self.app.post(
            BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)

        # purchase the product
        new_product = resp.get_json()
        stock_num = new_product["stock"]
        logging.debug(new_product)
        resp = self.app.put(
            "/products/{}/purchase".format(new_product["id"]),
            json=new_product,
            content_type=CONTENT_TYPE_JSON,
        )
        self.assertEqual(resp.status_code, status.HTTP_200_OK)
        purchased_product = resp.get_json()
        self.assertEqual(purchased_product["stock"], stock_num - 1)

    def test_purchase_product_out_of_stock(self):
        """Purchase a out of stock Product"""
        # create a product to purchase
        test_product = Product(name="Xiaomi", category="Phone", description="Test for purchase", price=999, stock=0)
        resp = self.app.post(
            BASE_URL, json=test_product.serialize(), content_type=CONTENT_TYPE_JSON
        )
        self.assertEqual(resp.status_code, status.HTTP_201_CREATED)

        # purchase the product
        new_product = resp.get_json()
        stock_num = new_product["stock"]
        logging.debug(new_product)
        resp = self.app.put(
            "/products/{}/purchase".format(new_product["id"]),
            json=new_product,
            content_type=CONTENT_TYPE_JSON,
        )
        self.assertEqual(resp.status_code, status.HTTP_409_CONFLICT)
